+++ app/Policies/Partner/CustomerPolicy.php
@@ -1,0 +1,100 @@
+<?php
+
+namespace App\Policies\Partner;
+
+use App\JsonApi\Proxies\CustomerPartner as Customer;
+use App\Models\User;
+use Illuminate\Auth\Access\HandlesAuthorization;
+
+class CustomerPolicy
+{
+    use HandlesAuthorization;
+
+    /**
+     * Determine whether the user can view any models.
+     *
+     * @param  \App\Models\User  $user
+     * @return \Illuminate\Auth\Access\Response|bool
+     */
+    public function viewAny(User $user)
+    {
+        return true;
+       // return $user->isA('partner');
+    }
+
+    /**
+     * Determine whether the user can view the model.
+     *
+     * @param  \App\Models\User  $user
+     * @param  Customer  $customer
+     * @return \Illuminate\Auth\Access\Response|bool
+     */
+    public function view(User $user, Customer  $customer)
+    {
+        return $user->isA('partner');
+    }
+
+    /**
+     * Determine whether the user can create models.
+     *
+     * @param  \App\Models\User  $user
+     * @return \Illuminate\Auth\Access\Response|bool
+     */
+    public function create(User $user)
+    {
+        return false;
+        // return !$user->isBlock;
+    }
+
+    /**
+     * Determine whether the user can update the model.
+     *
+     * @param  \App\Models\User  $user
+     * @param  Customer  $customer
+     * @return \Illuminate\Auth\Access\Response|bool
+     */
+    public function update(User $user, Customer  $customer)
+    {
+        return false;
+        // return $user->id == $company->admin_user_id || $user->isA('superadmin', "admin");
+    }
+
+    /**
+     * Determine whether the user can delete the model.
+     *
+     * @param  \App\Models\User  $user
+     * @param  Customer  $customer
+     * @return \Illuminate\Auth\Access\Response|bool
+     */
+    public function delete(User $user, Customer  $customer)
+    {
+        return false;
+        // return $user->id == $company->admin_user_id || $user->isA('superadmin', "admin");
+    }
+
+    /**
+     * Determine whether the user can restore the model.
+     *
+     * @param  \App\Models\User  $user
+     * @param  Customer  $customer
+     * @return \Illuminate\Auth\Access\Response|bool
+     */
+    public function restore(User $user, Customer  $customer)
+    {
+        return false;
+        //return $user->id == $company->admin_user_id || $user->isA('superadmin', "admin");
+    }
+
+    /**
+     * Determine whether the user can permanently delete the model.
+     *
+     * @param  \App\Models\User  $user
+     * @param  Customer  $customer
+     * @return \Illuminate\Auth\Access\Response|bool
+     */
+    public function forceDelete(User $user, Customer  $customer)
+    {
+        //
+        return false;
+    }
+}